rename socket package to sockets
